services/db_mv_refreshment/requirements.txt
psycopg2_binary
python-dotenv
python_dateutil
pytz
schedule
pandas
holidays
sqlalchemy
requests
